studio/requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.23.0
blinker==1.9.0
blockbuster==1.5.24; python_version >= "3.11"
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1; python_version >= "3.11"
colorama==0.4.6; sys_platform == "win32" or platform_system == "Windows"
cryptography==44.0.3
decorator==5.2.1
distro==1.9.0
et-xmlfile==2.0.0
executing==2.2.0
forbiddenfruit==0.1.4; implementation_name == "cpython" and python_version >= "3.11"
gitdb==4.0.12
gitpython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-rs==0.29.1; python_version >= "3.11"
jsonschema-specifications==2025.4.1
langchain-core==0.3.68
langchain-openai==0.3.27
langgraph==0.5.2
langgraph-api==0.2.86; python_version >= "3.11"
langgraph-checkpoint==2.1.0
langgraph-cli[inmem]==0.3.4
langgraph-prebuilt==0.5.2
langgraph-runtime-inmem==0.3.4; python_version >= "3.11"
langgraph-sdk==0.1.72
langsmith==0.4.5
lxml==6.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
msal==1.32.3
msal-extensions==1.3.1
narwhals==1.46.0
numpy==2.3.1
openai[azure]==1.95.1
openpyxl==3.1.5
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0; sys_platform != "win32" and sys_platform != "emscripten"
pillow==11.3.0
prompt-toolkit==3.0.51
protobuf==6.31.1
ptyprocess==0.7.0; sys_platform != "win32" and sys_platform != "emscripten"
pure-eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pygments==2.19.2
pyjwt[crypto]==2.10.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.1.3; python_version >= "3.11"
stack-data==0.6.3
starlette==0.47.1; python_version >= "3.11"
streamlit==1.46.1
structlog==25.4.0; python_version >= "3.11"
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.1; python_version >= "3.11"
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0; python_version >= "3.11"
watchdog==6.0.0; platform_system != "Darwin"
watchfiles==1.1.0; python_version >= "3.11"
wcwidth==0.2.13
xxhash==3.5.0
zstandard==0.23.0
